refactor: migrate usePackets to TanStack Query useInfiniteQuery
- Replace useState/useEffect with useInfiniteQuery for automatic caching
- Share cache between main window and pop-out window (fixes duplicate packets bug)
- Export PACKETS_QUERY_KEY and getPacketsQueryKey for external cache access
- Use refetchInterval for polling instead of manual setInterval
- Update tests to use QueryClientProvider wrapper
- Rename test file to .tsx for JSX support

fix: remove fake timer usage from usePackets tests

- Replace vi.advanceTimersByTimeAsync with real timer/waitFor
- Simplify rate limit tests to work with TanStack Query
- Fix canView false test to check packets instead of loading state
- All 23 tests now pass
