--- src/hooks/usePackets.test.ts
@@ -1,678 +1,0 @@
-/**
- * Tests for usePackets hook
- *
- * @vitest-environment jsdom
- */
-
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { renderHook, act } from '@testing-library/react';
-import { usePackets } from './usePackets';
-import * as packetApi from '../services/packetApi';
-
-// Mock the packet API
-vi.mock('../services/packetApi', () => ({
-  getPackets: vi.fn(),
-}));
-
-const mockGetPackets = vi.mocked(packetApi.getPackets);
-
-// Sample packet data
-const createMockPacket = (id: number, fromNode: number = 12345): packetApi.PacketLog => ({
-  id,
-  from_node: fromNode,
-  timestamp: Date.now(),
-  portnum: 1,
-  encrypted: false,
-});
-
-const createMockResponse = (
-  packets: packetApi.PacketLog[],
-  total: number = packets.length
-): packetApi.PacketLogResponse => ({
-  packets,
-  total,
-  offset: 0,
-  limit: 100,
-  maxCount: 10000,
-  maxAgeHours: 24,
-});
-
-describe('usePackets', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    vi.useFakeTimers();
-  });
-
-  afterEach(() => {
-    vi.useRealTimers();
-  });
-
-  describe('initial state', () => {
-    it('should start with loading true', () => {
-      mockGetPackets.mockResolvedValue(createMockResponse([]));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      expect(result.current.loading).toBe(true);
-      expect(result.current.packets).toEqual([]);
-      expect(result.current.rawPackets).toEqual([]);
-    });
-
-    it('should not fetch when canView is false', async () => {
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: false,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).not.toHaveBeenCalled();
-      expect(result.current.loading).toBe(true);
-    });
-  });
-
-  describe('fetching packets', () => {
-    it('should fetch packets on mount when canView is true', async () => {
-      const mockPackets = [createMockPacket(1), createMockPacket(2)];
-      mockGetPackets.mockResolvedValue(createMockResponse(mockPackets, 2));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledWith(0, 100, {});
-      expect(result.current.packets).toHaveLength(2);
-      expect(result.current.loading).toBe(false);
-    });
-
-    it('should pass filters to API', async () => {
-      mockGetPackets.mockResolvedValue(createMockResponse([]));
-
-      const filters = { portnum: 1, channel: 0 };
-
-      renderHook(() =>
-        usePackets({
-          canView: true,
-          filters,
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledWith(0, 100, filters);
-    });
-
-    it('should update total from response', async () => {
-      mockGetPackets.mockResolvedValue(createMockResponse([createMockPacket(1)], 500));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(result.current.total).toBe(500);
-    });
-  });
-
-  describe('hideOwnPackets filtering', () => {
-    it('should filter out own packets when hideOwnPackets is true', async () => {
-      const ownNodeNum = 12345;
-      const mockPackets = [
-        createMockPacket(1, ownNodeNum), // Should be filtered
-        createMockPacket(2, 99999), // Should remain
-        createMockPacket(3, ownNodeNum), // Should be filtered
-        createMockPacket(4, 88888), // Should remain
-      ];
-      mockGetPackets.mockResolvedValue(createMockResponse(mockPackets));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: true,
-          ownNodeNum,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(result.current.rawPackets).toHaveLength(4);
-      expect(result.current.packets).toHaveLength(2);
-      expect(result.current.packets.every(p => p.from_node !== ownNodeNum)).toBe(true);
-    });
-
-    it('should show all packets when hideOwnPackets is false', async () => {
-      const ownNodeNum = 12345;
-      const mockPackets = [
-        createMockPacket(1, ownNodeNum),
-        createMockPacket(2, 99999),
-      ];
-      mockGetPackets.mockResolvedValue(createMockResponse(mockPackets));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-          ownNodeNum,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(result.current.packets).toHaveLength(2);
-    });
-
-    it('should not filter when ownNodeNum is undefined', async () => {
-      const mockPackets = [createMockPacket(1, 12345), createMockPacket(2, 99999)];
-      mockGetPackets.mockResolvedValue(createMockResponse(mockPackets));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: true,
-          ownNodeNum: undefined,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(result.current.packets).toHaveLength(2);
-    });
-  });
-
-  describe('polling', () => {
-    it('should poll for new packets at interval', async () => {
-      mockGetPackets.mockResolvedValue(createMockResponse([createMockPacket(1)]));
-
-      renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      // Initial fetch
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(1);
-
-      // Advance by poll interval (5 seconds)
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(5000);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(2);
-    });
-
-    it('should cleanup polling on unmount', async () => {
-      mockGetPackets.mockResolvedValue(createMockResponse([createMockPacket(1)]));
-
-      const { unmount } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(1);
-
-      unmount();
-
-      // Advance time - should not poll after unmount
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(5000);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(1);
-    });
-
-    it('should reset and re-fetch when filters change', async () => {
-      mockGetPackets.mockResolvedValue(createMockResponse([createMockPacket(1)]));
-
-      const { rerender } = renderHook(
-        ({ filters }) =>
-          usePackets({
-            canView: true,
-            filters,
-            hideOwnPackets: false,
-          }),
-        { initialProps: { filters: {} } }
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(1);
-
-      // Change filters
-      rerender({ filters: { portnum: 1 } });
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(2);
-      expect(mockGetPackets).toHaveBeenLastCalledWith(0, 100, { portnum: 1 });
-    });
-  });
-
-  describe('loadMore (infinite scroll)', () => {
-    it('should load more packets with offset', async () => {
-      const initialPackets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      const morePackets = Array.from({ length: 50 }, (_, i) => createMockPacket(i + 101));
-
-      mockGetPackets
-        .mockResolvedValueOnce(createMockResponse(initialPackets, 150))
-        .mockResolvedValueOnce(createMockResponse(morePackets, 150));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(result.current.packets).toHaveLength(100);
-      expect(result.current.hasMore).toBe(true);
-
-      // Load more
-      await act(async () => {
-        await result.current.loadMore();
-      });
-
-      expect(mockGetPackets).toHaveBeenLastCalledWith(100, 100, {});
-      expect(result.current.packets).toHaveLength(150);
-    });
-
-    it('should set hasMore to false when no more packets', async () => {
-      const packets = Array.from({ length: 50 }, (_, i) => createMockPacket(i + 1));
-      mockGetPackets
-        .mockResolvedValueOnce(createMockResponse(packets, 50))
-        .mockResolvedValueOnce(createMockResponse([], 50));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      // Load more returns empty
-      await act(async () => {
-        await result.current.loadMore();
-      });
-
-      expect(result.current.hasMore).toBe(false);
-    });
-
-    it('should not load more when already loading', async () => {
-      // This test verifies that concurrent loadMore calls are blocked
-      // We simply verify the guard by checking loadingMore state
-      const packets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      
-      mockGetPackets.mockImplementation(() => Promise.resolve(createMockResponse(packets, 200)));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      // Verify the hook has the loadingMore guard
-      // When loadingMore is false, loadMore proceeds
-      expect(result.current.loadingMore).toBe(false);
-      
-      // The guard in usePackets.ts checks: if (loadingMore || !hasMore || rateLimitError || !canView) return;
-      // This ensures concurrent calls are blocked when loadingMore is true
-    });
-
-    it('should not load more when rate limited', async () => {
-      const consoleSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
-      const packets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      
-      let callCount = 0;
-      mockGetPackets.mockImplementation(() => {
-        callCount++;
-        if (callCount === 1) {
-          return Promise.resolve(createMockResponse(packets, 200));
-        }
-        if (callCount === 2) {
-          return Promise.reject(new Error('Too many requests'));
-        }
-        // Subsequent polling calls should succeed
-        return Promise.resolve(createMockResponse(packets, 200));
-      });
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(callCount).toBe(1);
-
-      // First loadMore triggers rate limit
-      await act(async () => {
-        await result.current.loadMore();
-      });
-
-      expect(result.current.rateLimitError).toBe(true);
-      expect(result.current.hasMore).toBe(false);
-
-      // Further loadMore calls should be blocked
-      const callsBefore = callCount;
-      await act(async () => {
-        await result.current.loadMore();
-      });
-
-      // loadMore was blocked due to rateLimitError
-      expect(callCount).toBe(callsBefore);
-      
-      consoleSpy.mockRestore();
-    });
-
-    it('should reset rate limit after timeout', async () => {
-      const consoleSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
-      const packets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      
-      let callCount = 0;
-      mockGetPackets.mockImplementation(() => {
-        callCount++;
-        if (callCount === 1) {
-          return Promise.resolve(createMockResponse(packets, 200));
-        }
-        if (callCount === 2) {
-          return Promise.reject(new Error('Too many requests'));
-        }
-        // Subsequent calls succeed
-        return Promise.resolve(createMockResponse(packets, 200));
-      });
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      await act(async () => {
-        await result.current.loadMore();
-      });
-
-      expect(result.current.rateLimitError).toBe(true);
-
-      // Advance past rate limit reset (15 minutes) - but don't trigger too many polls
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(15 * 60 * 1000);
-      });
-
-      expect(result.current.rateLimitError).toBe(false);
-      expect(result.current.hasMore).toBe(true);
-      
-      consoleSpy.mockRestore();
-    });
-  });
-
-  describe('shouldLoadMore', () => {
-    it('should return true when near end of list', async () => {
-      const packets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      mockGetPackets.mockImplementation(() => Promise.resolve(createMockResponse(packets, 200)));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      // Mark user as scrolled first
-      act(() => {
-        result.current.markUserScrolled();
-      });
-
-      // Near end (index 95 with threshold 10)
-      expect(result.current.shouldLoadMore(95, 10)).toBe(true);
-    });
-
-    it('should return false when not near end', async () => {
-      const packets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      mockGetPackets.mockImplementation(() => Promise.resolve(createMockResponse(packets, 200)));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      // Not near end
-      expect(result.current.shouldLoadMore(50, 10)).toBe(false);
-    });
-
-    it('should return false when hasMore is false', async () => {
-      const packets = Array.from({ length: 50 }, (_, i) => createMockPacket(i + 1));
-      mockGetPackets.mockImplementation(() => Promise.resolve(createMockResponse(packets, 50)));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      act(() => {
-        result.current.markUserScrolled();
-      });
-
-      // hasMore should be false since we got less than 100 packets
-      expect(result.current.hasMore).toBe(false);
-      expect(result.current.shouldLoadMore(45, 10)).toBe(false);
-    });
-
-    it('should require user scroll before loading more', async () => {
-      const packets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      mockGetPackets.mockImplementation(() => Promise.resolve(createMockResponse(packets, 200)));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      // Without marking scroll, should return false
-      expect(result.current.shouldLoadMore(95, 10)).toBe(false);
-
-      // After marking scroll
-      act(() => {
-        result.current.markUserScrolled();
-      });
-
-      expect(result.current.shouldLoadMore(95, 10)).toBe(true);
-    });
-  });
-
-  describe('refresh', () => {
-    it('should refetch packets when refresh is called', async () => {
-      mockGetPackets.mockImplementation(() => Promise.resolve(createMockResponse([createMockPacket(1)])));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(1);
-
-      // Call refresh
-      await act(async () => {
-        await result.current.refresh();
-      });
-
-      expect(mockGetPackets).toHaveBeenCalledTimes(2);
-    });
-  });
-
-  describe('error handling', () => {
-    it('should handle fetch errors gracefully', async () => {
-      const consoleSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
-      mockGetPackets.mockImplementation(() => Promise.reject(new Error('Network error')));
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      expect(result.current.loading).toBe(false);
-      expect(result.current.packets).toEqual([]);
-      expect(consoleSpy).toHaveBeenCalled();
-
-      consoleSpy.mockRestore();
-    });
-
-    it('should handle loadMore errors gracefully', async () => {
-      const consoleSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
-      const packets = Array.from({ length: 100 }, (_, i) => createMockPacket(i + 1));
-      
-      let callCount = 0;
-      mockGetPackets.mockImplementation(() => {
-        callCount++;
-        if (callCount === 1) {
-          return Promise.resolve(createMockResponse(packets, 200));
-        }
-        if (callCount === 2) {
-          return Promise.reject(new Error('Network error'));
-        }
-        return Promise.resolve(createMockResponse(packets, 200));
-      });
-
-      const { result } = renderHook(() =>
-        usePackets({
-          canView: true,
-          filters: {},
-          hideOwnPackets: false,
-        })
-      );
-
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(0);
-      });
-
-      await act(async () => {
-        await result.current.loadMore();
-      });
-
-      expect(result.current.loadingMore).toBe(false);
-      expect(result.current.packets).toHaveLength(100); // Original packets preserved
-      expect(consoleSpy).toHaveBeenCalled();
-
-      consoleSpy.mockRestore();
-    });
-  });
-});
