+++ src/hooks/usePackets.ts
@@ -1,20 +1,38 @@
 /**
- * Hook for managing packet data fetching and state
+ * Hook for managing packet data fetching and state using TanStack Query
  *
  * Provides packet data management for PacketMonitorPanel including:
- * - Initial fetch and polling
- * - Infinite scroll/load more
+ * - Initial fetch and polling with automatic cache management
+ * - Infinite scroll/load more with useInfiniteQuery
  * - Filtering (server-side and client-side)
  * - Rate limit handling
+ *
+ * Uses TanStack Query for:
+ * - Automatic request deduplication
+ * - Cache sharing between components (fixes pop-out window duplicate packets)
+ * - Stale-while-revalidate pattern
  */
 
-import { useState, useEffect, useCallback, useRef, useMemo } from 'react';
+import { useCallback, useRef, useMemo, useState, useEffect } from 'react';
+import { useInfiniteQuery, useQueryClient } from '@tanstack/react-query';
 import { PacketLog, PacketFilters } from '../types/packet';
 import { getPackets } from '../services/packetApi';
 
 // Constants
 const PACKET_FETCH_LIMIT = 100;
 const POLL_INTERVAL_MS = 5000;
+
+/**
+ * Query key for packets - includes filters for proper cache isolation
+ */
+export const PACKETS_QUERY_KEY = ['packets'] as const;
+
+/**
+ * Build the full query key including filters
+ */
+export function getPacketsQueryKey(filters: PacketFilters) {
+  return [...PACKETS_QUERY_KEY, filters] as const;
+}
 
 interface UsePacketsOptions {
   /** Whether the user has permission to view packets */
@@ -55,6 +73,10 @@ interface UsePacketsResult {
 /**
  * Hook to manage packet data fetching and state
  *
+ * Uses TanStack Query's useInfiniteQuery for automatic caching and
+ * request deduplication. This fixes issues with duplicate packets
+ * appearing in pop-out windows since cache is shared.
+ *
  * @param options - Configuration options
  * @returns Packet data and controls
  *
@@ -74,23 +96,59 @@ interface UsePacketsResult {
  * });
  * ```
  */
-export function usePackets({
-  canView,
-  filters,
-  hideOwnPackets,
-  ownNodeNum,
-}: UsePacketsOptions): UsePacketsResult {
-  const [rawPackets, setRawPackets] = useState<PacketLog[]>([]);
-  const [total, setTotal] = useState(0);
-  const [loading, setLoading] = useState(true);
-  const [loadingMore, setLoadingMore] = useState(false);
-  const [hasMore, setHasMore] = useState(true);
-  const [rateLimitError, setRateLimitError] = useState(false);
+export function usePackets({ canView, filters, hideOwnPackets, ownNodeNum }: UsePacketsOptions): UsePacketsResult {
+  const queryClient = useQueryClient();
 
-  const pollIntervalRef = useRef<ReturnType<typeof setInterval> | null>(null);
+  // Rate limit state (not handled by React Query)
+  const [rateLimitError, setRateLimitError] = useState(false);
   const rateLimitResetTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+
+  // Scroll tracking refs
   const lastLoadedRawLengthRef = useRef<number>(0);
   const userHasScrolledRef = useRef<boolean>(false);
+
+  // Query key for this filter combination
+  const queryKey = useMemo(() => getPacketsQueryKey(filters), [filters]);
+
+  // Use infinite query for paginated data with polling
+  const { data, fetchNextPage, hasNextPage, isFetchingNextPage, isLoading, refetch } = useInfiniteQuery({
+    queryKey,
+    queryFn: async ({ pageParam = 0 }) => {
+      const response = await getPackets(pageParam, PACKET_FETCH_LIMIT, filters);
+      return response;
+    },
+    initialPageParam: 0,
+    getNextPageParam: (lastPage, allPages) => {
+      // Calculate total loaded packets across all pages
+      const totalLoaded = allPages.reduce((sum, page) => sum + page.packets.length, 0);
+
+      // If last page has fewer packets than limit, there are no more
+      if (lastPage.packets.length < PACKET_FETCH_LIMIT) {
+        return undefined;
+      }
+
+      // Return the offset for the next page
+      return totalLoaded;
+    },
+    enabled: canView,
+    refetchInterval: POLL_INTERVAL_MS,
+    refetchIntervalInBackground: false,
+    staleTime: POLL_INTERVAL_MS - 1000, // Consider stale just before next poll
+    gcTime: 5 * 60 * 1000, // Keep in cache for 5 minutes
+    refetchOnWindowFocus: false,
+  });
+
+  // Flatten all pages into raw packets array
+  const rawPackets = useMemo(() => {
+    if (!data?.pages) return [];
+    return data.pages.flatMap(page => page.packets);
+  }, [data?.pages]);
+
+  // Get total from the most recent page response
+  const total = useMemo(() => {
+    if (!data?.pages || data.pages.length === 0) return 0;
+    return data.pages[data.pages.length - 1].total;
+  }, [data?.pages]);
 
   // Apply client-side "Hide Own Packets" filter
   const packets = useMemo(() => {
@@ -100,70 +158,25 @@ export function usePackets({
     return rawPackets;
   }, [rawPackets, hideOwnPackets, ownNodeNum]);
 
-  // Reset scroll tracking when raw packets are cleared
+  // Reset scroll tracking when filters change (query is reset)
   useEffect(() => {
-    if (rawPackets.length === 0) {
-      userHasScrolledRef.current = false;
-      lastLoadedRawLengthRef.current = 0;
-    }
-  }, [rawPackets.length]);
-
-  // Fetch packets (initial load or refresh from polling)
-  const fetchPackets = useCallback(async () => {
-    if (!canView) return;
-
-    try {
-      const currentPacketCount = rawPackets.length;
-      const isInitialLoad = currentPacketCount === 0;
-
-      const response = await getPackets(0, PACKET_FETCH_LIMIT, filters);
-
-      // Only update if initial load OR if there are new packets
-      if (isInitialLoad || response.packets[0]?.id !== rawPackets[0]?.id) {
-        if (!isInitialLoad && currentPacketCount > PACKET_FETCH_LIMIT) {
-          // Preserve existing packets beyond the first batch when polling
-          const newPacketIds = new Set(response.packets.map(p => p.id));
-          const oldPacketsWithoutDuplicates = rawPackets.filter(p => !newPacketIds.has(p.id));
-          setRawPackets([...response.packets, ...oldPacketsWithoutDuplicates]);
-        } else {
-          setRawPackets(response.packets);
-          setHasMore(response.packets.length >= PACKET_FETCH_LIMIT);
-        }
-        setTotal(response.total);
-      }
-
-      setLoading(false);
-    } catch (error) {
-      console.error('Failed to fetch packets:', error);
-      setLoading(false);
-    }
-  }, [canView, filters, rawPackets]);
+    userHasScrolledRef.current = false;
+    lastLoadedRawLengthRef.current = 0;
+  }, [queryKey]);
 
   // Load more packets (infinite scroll)
   const loadMore = useCallback(async () => {
-    if (loadingMore || !hasMore || rateLimitError || !canView) return;
+    if (isFetchingNextPage || !hasNextPage || rateLimitError || !canView) return;
 
-    setLoadingMore(true);
     try {
-      const response = await getPackets(rawPackets.length, PACKET_FETCH_LIMIT, filters);
-
-      if (response.packets.length === 0) {
-        setHasMore(false);
-      } else {
-        setRawPackets(prev => {
-          const newLength = prev.length + response.packets.length;
-          lastLoadedRawLengthRef.current = newLength;
-          return [...prev, ...response.packets];
-        });
-        setTotal(response.total);
-      }
+      await fetchNextPage();
+      lastLoadedRawLengthRef.current = rawPackets.length + PACKET_FETCH_LIMIT;
     } catch (error) {
       console.error('Failed to load more packets:', error);
 
       // Check for rate limit error
       if (error instanceof Error && error.message.includes('Too many requests')) {
         setRateLimitError(true);
-        setHasMore(false);
 
         if (rateLimitResetTimerRef.current) {
           clearTimeout(rateLimitResetTimerRef.current);
@@ -172,13 +185,17 @@ export function usePackets({
         // Reset after 15 minutes
         rateLimitResetTimerRef.current = setTimeout(() => {
           setRateLimitError(false);
-          setHasMore(true);
         }, 15 * 60 * 1000);
       }
-    } finally {
-      setLoadingMore(false);
     }
-  }, [loadingMore, hasMore, rateLimitError, canView, rawPackets.length, filters]);
+  }, [isFetchingNextPage, hasNextPage, rateLimitError, canView, fetchNextPage, rawPackets.length]);
+
+  // Refresh packets (invalidate cache and refetch)
+  const refresh = useCallback(async () => {
+    // Invalidate and refetch
+    await queryClient.invalidateQueries({ queryKey });
+    await refetch();
+  }, [queryClient, queryKey, refetch]);
 
   // Mark that user has scrolled (enables infinite scroll loading)
   const markUserScrolled = useCallback(() => {
@@ -186,49 +203,35 @@ export function usePackets({
   }, []);
 
   // Check if should load more based on scroll position
-  const shouldLoadMore = useCallback((lastVisibleIndex: number, threshold = 10): boolean => {
-    if (lastVisibleIndex < 0) return false;
+  const shouldLoadMore = useCallback(
+    (lastVisibleIndex: number, threshold = 10): boolean => {
+      if (lastVisibleIndex < 0) return false;
 
-    // Don't load if we have raw packets but no filtered packets
-    if (packets.length === 0 && rawPackets.length > 0) return false;
+      // Don't load if we have raw packets but no filtered packets
+      if (packets.length === 0 && rawPackets.length > 0) return false;
 
-    const nearEnd = lastVisibleIndex >= packets.length - threshold;
-    const alreadyLoadedFromHere = rawPackets.length > 0 && rawPackets.length === lastLoadedRawLengthRef.current;
-    const enoughItemsDisplayed = packets.length >= threshold;
-    const shouldRequireScroll = enoughItemsDisplayed && !userHasScrolledRef.current;
+      const nearEnd = lastVisibleIndex >= packets.length - threshold;
+      const alreadyLoadedFromHere = rawPackets.length > 0 && rawPackets.length === lastLoadedRawLengthRef.current;
+      const enoughItemsDisplayed = packets.length >= threshold;
+      const shouldRequireScroll = enoughItemsDisplayed && !userHasScrolledRef.current;
+      const hasMorePages = hasNextPage ?? true;
 
-    return nearEnd && hasMore && !loadingMore && !alreadyLoadedFromHere && !shouldRequireScroll && canView;
-  }, [packets.length, rawPackets.length, hasMore, loadingMore, canView]);
+      return (
+        nearEnd &&
+        hasMorePages &&
+        !isFetchingNextPage &&
+        !alreadyLoadedFromHere &&
+        !shouldRequireScroll &&
+        canView &&
+        !rateLimitError
+      );
+    },
+    [packets.length, rawPackets.length, hasNextPage, isFetchingNextPage, canView, rateLimitError]
+  );
 
-  // Initial fetch and polling
-  useEffect(() => {
-    if (!canView) return;
-
-    // Reset state when filters change
-    setRawPackets([]);
-    setLoading(true);
-    setHasMore(true);
-    lastLoadedRawLengthRef.current = 0;
-    userHasScrolledRef.current = false;
-
-    fetchPackets();
-
-    pollIntervalRef.current = setInterval(fetchPackets, POLL_INTERVAL_MS);
-
-    return () => {
-      if (pollIntervalRef.current) {
-        clearInterval(pollIntervalRef.current);
-      }
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [canView, JSON.stringify(filters)]); // fetchPackets not in deps to avoid infinite loop
-
-  // Cleanup on unmount
+  // Cleanup rate limit timer on unmount
   useEffect(() => {
     return () => {
-      if (pollIntervalRef.current) {
-        clearInterval(pollIntervalRef.current);
-      }
       if (rateLimitResetTimerRef.current) {
         clearTimeout(rateLimitResetTimerRef.current);
       }
@@ -239,12 +242,12 @@ export function usePackets({
     packets,
     rawPackets,
     total,
-    loading,
-    loadingMore,
-    hasMore,
+    loading: isLoading,
+    loadingMore: isFetchingNextPage,
+    hasMore: hasNextPage ?? true,
     rateLimitError,
     loadMore,
-    refresh: fetchPackets,
+    refresh,
     markUserScrolled,
     shouldLoadMore,
   };
